backend/requirements.txt
Django==5.0.2
djangorestframework==3.15.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
Pillow==10.2.0
django-storages==1.14.2
boto3==1.34.34
celery==5.3.4
redis==5.0.1
django-filter==23.5
django-extensions==3.2.3
